Add wallet tests

+++ maverics/src/main/kotlin/com/example/services/AddPlatformCharge.kt
@@ -1,9 +1,8 @@
 package com.example.services
 
 import com.example.constants.totalPlatformFees
-import com.example.controller.PlatformFees
 import java.math.BigInteger
 
-fun addPlatformCharge(platformCharge: Long){
-    totalPlatformFees += BigInteger.valueOf(platformCharge)
+fun addPlatformCharge(platformCharge: BigInteger) {
+    totalPlatformFees += platformCharge
 }
